+++ README.md
@@ -1,3 +1,3 @@
-# embulk-util-web_api
+# embulk-base-restclient
 
-Common Libraries for Web API based Embulk plugins.  Please see sample code: https://github.com/muga/embulk-util-web_api/blob/master/src/test/java/org/embulk/input/shopify/ShopifyInputPlugin.java
+Base classes for Embulk plugins working as REST clients. Please see sample code: https://github.com/embulk/embulk-base-restclient/blob/master/src/test/java/org/embulk/input/shopify/ShopifyInputPlugin.java
