+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'embulk-util-web_api'
+rootProject.name = 'embulk-base-restclient'
