+++ src/main/java/org/embulk/base/restclient/WebApiPluginTask.java
@@ -1,4 +1,4 @@
-package org.embulk.util.web_api;
+package org.embulk.base.restclient;
 
 import org.embulk.config.Config;
 import org.embulk.config.ConfigDefault;
