+++ src/main/java/org/embulk/base/restclient/writer/JsonColumnWriter.java
@@ -1,10 +1,12 @@
-package org.embulk.util.web_api.writer;
+package org.embulk.base.restclient.writer;
 
 import com.fasterxml.jackson.databind.JsonNode;
+
 import org.embulk.spi.Column;
 import org.embulk.spi.PageBuilder;
 import org.embulk.spi.json.JsonParser;
-import org.embulk.util.web_api.writer.SchemaWriterFactory.WebApiColumnOption;
+
+import org.embulk.base.restclient.writer.SchemaWriterFactory.WebApiColumnOption;
 
 public class JsonColumnWriter
         extends AbstractColumnWriter
