+++ src/main/java/org/embulk/base/restclient/writer/StringColumnWriter.java
@@ -1,14 +1,16 @@
-package org.embulk.util.web_api.writer;
+package org.embulk.base.restclient.writer;
 
 import com.fasterxml.jackson.databind.JsonNode;
+
 import org.embulk.spi.Column;
 import org.embulk.spi.PageBuilder;
-import org.embulk.util.web_api.writer.SchemaWriterFactory.WebApiColumnOption;
 
-public class DoubleColumnWriter
+import org.embulk.base.restclient.writer.SchemaWriterFactory.WebApiColumnOption;
+
+public class StringColumnWriter
         extends AbstractColumnWriter
 {
-    public DoubleColumnWriter(Column column, WebApiColumnOption option)
+    public StringColumnWriter(Column column, WebApiColumnOption option)
     {
         super(column, option);
     }
@@ -20,7 +22,7 @@ public class DoubleColumnWriter
             to.setNull(column);
         }
         else {
-            to.setDouble(column, v.doubleValue());
+            to.setString(column, v.asText());
         }
     }
 }
